[FE]: [REFACTOR] #20 Tools 리팩토링
- buttonData 분리
