+++ client/src/components/organisms/Tools/Tools.tsx
@@ -1,20 +1,10 @@
 import React, { useState, useReducer } from 'react';
-import styled from 'styled-components';
+import styled, { keyframes } from 'styled-components';
 import { shallowEqual, useDispatch, useSelector } from 'react-redux';
-import {
-  BsFillSkipStartFill,
-  BsFillSkipEndFill,
-  BsFillPlayFill,
-  BsFillPauseFill,
-  BsAspectRatio,
-} from 'react-icons/bs';
-import { RiScissorsLine } from 'react-icons/ri';
-import { MdScreenRotation } from 'react-icons/md';
 
 import webglController from '@/webgl/webglController';
 import ButtonGroup from '@/components/molecules/ButtonGroup';
 import UploadArea from '@/components/molecules/UploadArea';
-import size from '@/theme/sizes';
 import video from '@/video';
 import {
   play as playAction,
@@ -24,14 +14,33 @@ import {
 import { getStartEnd } from '@/store/selectors';
 import { cropStart, cropCancel, cropConfirm } from '@/store/actionTypes';
 
-import reducer, { initialData, ButtonData, ButtonDataAction } from './reducer';
+import reducer, { initialData, ButtonDataAction } from './reducer';
+import {
+  getEditToolData,
+  getSubEditToolsData,
+  getVideoToolsData,
+} from './buttonData';
+
+const UP = 'up';
+const DOWN = 'down';
 
 const StyledDiv = styled.div`
+  position: relative;
   display: flex;
   justify-content: space-between;
   align-items: flex-end;
   padding: 1rem;
-  height: 10rem;
+`;
+
+const slide = keyframes`
+  from {
+    transform: translate(0, 50px);
+    opacity: 0;
+  }
+  to {
+    transform: translate(0, 0);
+    opacity: 1;
+  }
 `;
 
 const StyledEditToolDiv = styled.div`
@@ -43,82 +52,25 @@ const StyledEditToolDiv = styled.div`
 const EditTool = styled(ButtonGroup)`
   width: 20rem;
 `;
-const SubEditTool = styled(ButtonGroup)`
+
+const WrapperDiv = styled.div`
   width: 25rem;
+  display: flex;
+  justify-content: center;
+  animation: ${slide} 0.5s -0.1s ease-out;
 `;
+
+const SubEditTool = styled(ButtonGroup)`
+  width: 100%;
+`;
+
 const VideoTool = styled(ButtonGroup)``;
 
-interface button {
-  onClick: () => void;
-  message: string;
-  type: 'default' | 'transparent';
-  children: React.ReactChild;
+interface props {
+  setEdit: Function;
 }
 
-const getVideoToolsData = (
-  backwardVideo: () => void,
-  playPauseVideo: () => void,
-  forwardVideo: () => void,
-  play: boolean
-): button[] => [
-  {
-    onClick: backwardVideo,
-    message: '',
-    type: 'transparent',
-    children: <BsFillSkipStartFill size={size.BIG_ICON_SIZE} />,
-  },
-  {
-    onClick: playPauseVideo,
-    message: '',
-    type: 'transparent',
-    children: play ? (
-      <BsFillPlayFill size={size.BIG_ICON_SIZE} />
-    ) : (
-      <BsFillPauseFill size={size.BIG_ICON_SIZE} />
-    ),
-  },
-  {
-    onClick: forwardVideo,
-    message: '',
-    type: 'transparent',
-    children: <BsFillSkipEndFill size={size.BIG_ICON_SIZE} />,
-  },
-];
-
-const getEditToolData = (
-  rotateReverse: () => void,
-  ratio: () => void,
-  crop: () => void
-): button[] => [
-  {
-    onClick: rotateReverse,
-    message: '회전 / 반전',
-    type: 'transparent',
-    children: <MdScreenRotation size={size.ICON_SIZE} />,
-  },
-  {
-    onClick: ratio,
-    message: '비율',
-    type: 'transparent',
-    children: <BsAspectRatio size={size.ICON_SIZE} />,
-  },
-  {
-    onClick: crop,
-    message: '자르기',
-    type: 'transparent',
-    children: <RiScissorsLine size={size.ICON_SIZE} />,
-  },
-];
-
-const getSubEditToolsData = (buttonData: ButtonData): button[] =>
-  [...Array(buttonData.onClicks?.length)].map((_, idx) => ({
-    onClick: buttonData.onClicks[idx],
-    message: buttonData.messages[idx],
-    type: buttonData.type,
-    children: buttonData.childrens[idx],
-  }));
-
-const Tools: React.FC = () => {
+const Tools: React.FC<props> = ({ setEdit }) => {
   const [play, setPlay] = useState(true); // Fix 스토어로 등록
   const dispatch = useDispatch();
   const [toolType, setToolType] = useState(null);
@@ -148,7 +100,6 @@ const Tools: React.FC = () => {
       video.pause();
       dispatch(pause());
     }
-
     setPlay(!play);
   };
 
@@ -170,6 +121,21 @@ const Tools: React.FC = () => {
     }
   };
 
+  const closeSubtool = () => {
+    setEdit(DOWN);
+    setToolType(null);
+    dispatchButtonData({ type: null });
+  };
+
+  const openSubtool = (
+    type: 'videoEffect' | 'ratio' | 'crop',
+    payload: (() => void)[]
+  ) => {
+    setEdit(UP);
+    setToolType(type);
+    dispatchButtonData({ type, payload });
+  };
+
   const handleRotateLeft90Degree = () => webglController.rotateLeft90Degree();
   const handleRotateRight90Degree = () => webglController.rotateRight90Degree();
   const handleReverseUpsideDown = () => webglController.reverseUpsideDown();
@@ -181,56 +147,37 @@ const Tools: React.FC = () => {
     handleReverseSideToSide,
   ];
 
-  const handleRotateReverse = () => {
-    if (toolType === 'videoEffect') {
-      setToolType(null);
-      dispatchButtonData({ type: null });
-    } else {
-      setToolType('videoEffect');
-      dispatchButtonData({
-        type: 'videoEffect',
-        payload: rotateReverseMethods,
-      });
-    }
-  };
-
   const handleRatioEnlarge = () => webglController.enlarge();
   const handleRatioReduce = () => webglController.reduce();
   const ratioMethods = [handleRatioEnlarge, handleRatioReduce];
 
-  const handleRatio = () => {
-    if (toolType === 'ratio') {
-      setToolType(null);
-      dispatchButtonData({ type: null });
-    } else {
-      setToolType('ratio');
-      dispatchButtonData({ type: 'ratio', payload: ratioMethods });
-    }
-  };
-
-  const handleCropInsert = () => webglController.enlarge();
+  const handleCropInsert = () => null;
   const handleCropConfirm = () => {
     dispatch(cropConfirm());
-    setToolType(null);
-    dispatchButtonData({ type: null });
+    closeSubtool();
     dispatch(cropCancel());
   };
   const handleCropCancel = () => {
-    setToolType(null);
-    dispatchButtonData({ type: null });
+    closeSubtool();
     dispatch(cropCancel());
   };
   const cropMethods = [handleCropInsert, handleCropConfirm, handleCropCancel];
 
+  const handleRotateReverse = () =>
+    toolType === 'videoEffect'
+      ? closeSubtool()
+      : openSubtool('videoEffect', rotateReverseMethods);
+
+  const handleRatio = () =>
+    toolType === 'ratio' ? closeSubtool() : openSubtool('ratio', ratioMethods);
+
   const handleCrop = () => {
     if (toolType === 'crop') {
       dispatch(cropCancel());
-      setToolType(null);
-      dispatchButtonData({ type: null });
+      closeSubtool();
     } else {
+      openSubtool('crop', cropMethods);
       dispatch(cropStart());
-      setToolType('crop');
-      dispatchButtonData({ type: 'crop', payload: cropMethods });
     }
   };
 
@@ -245,7 +192,11 @@ const Tools: React.FC = () => {
         )}
       />
       <StyledEditToolDiv>
-        <SubEditTool buttonData={getSubEditToolsData(buttonData)} />
+        {toolType && (
+          <WrapperDiv>
+            <SubEditTool buttonData={getSubEditToolsData(buttonData)} />
+          </WrapperDiv>
+        )}
         <EditTool
           buttonData={getEditToolData(
             handleRotateReverse,
